🎉Source Zenloop: migrate to lowcode (#19624)
* Migrate to lowcode

* Updated PR

* Updated configured catalog

* Updated to review

* Updated after review

* Add expected records

* Updated docs

* Updated docs

* Fix to linter check

* auto-bump connector version
