+++ airbyte-config/init/src/main/resources/seed/source_specs.yaml
@@ -16433,7 +16433,7 @@
     supportsNormalization: false
     supportsDBT: false
     supported_destination_sync_modes: []
-- dockerImage: "airbyte/source-zenloop:0.1.3"
+- dockerImage: "airbyte/source-zenloop:0.1.4"
   spec:
     documentationUrl: "https://docs.airbyte.com/integrations/sources/zenloop"
     connectionSpecification:
