+++ airbyte-integrations/connectors/source-zenloop/Dockerfile
@@ -34,5 +34,5 @@ COPY source_zenloop ./source_zenloop
 ENV AIRBYTE_ENTRYPOINT "python /airbyte/integration_code/main.py"
 ENTRYPOINT ["python", "/airbyte/integration_code/main.py"]
 
-LABEL io.airbyte.version=0.1.3
+LABEL io.airbyte.version=0.1.4
 LABEL io.airbyte.name=airbyte/source-zenloop
