+++ airbyte-integrations/connectors/source-zenloop/acceptance-test-config.yml
@@ -13,31 +13,16 @@ tests:
     - config_path: "secrets/config.json"
   basic_read:
     - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_answers.json"
-      empty_streams: []
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_surveys.json"
-      empty_streams: []
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_answers_survey_group.json"
-      empty_streams: []
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_survey_groups.json"
-      empty_streams: []
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_properties.json"
-      empty_streams: []
+      configured_catalog_path: "integration_tests/configured_catalog.json"
+      expect_records:
+        path: "integration_tests/expected_records.txt"
+        extra_fields: no
+        exact_order: no
+        extra_records: yes
   incremental:
     - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_answers.json"
-      future_state_path: "integration_tests/abnormal_state.json"
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_answers_survey_group.json"
+      configured_catalog_path: "integration_tests/configured_catalog.json"
       future_state_path: "integration_tests/abnormal_state.json"
   full_refresh:
     - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_survey_groups.json"
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_surveys.json"
-    - config_path: "secrets/config.json"
-      configured_catalog_path: "integration_tests/configured_catalog_properties.json"
+      configured_catalog_path: "integration_tests/configured_catalog.json"
