+++ airbyte-integrations/connectors/source-zenloop/source_zenloop/zenloop.yaml
@@ -1,0 +1,149 @@
+version: "0.1.0"
+
+definitions:
+  selector:
+    type: RecordSelector
+    extractor:
+      type: DpathExtractor
+      field_pointer: ["{{ options['data_field'] }}"]
+  requester:
+    type: HttpRequester
+    name: "{{ options['name'] }}"
+    http_method: "GET"
+    authenticator:
+      type: BearerAuthenticator
+      api_token: "{{ config['api_token'] }}"
+  retriever:
+    type: SimpleRetriever
+    $options:
+      url_base: "https://api.zenloop.com/v1/"
+    name: "{{ options['name'] }}"
+    record_selector:
+      $ref: "*ref(definitions.selector)"
+    paginator:
+      type: DefaultPaginator
+      pagination_strategy:
+        type: PageIncrement
+        page_size: 50
+        start_from_page: 1
+      page_size_option:
+        field_name: "per_page"
+        inject_into: "request_parameter"
+      page_token_option:
+        inject_into: "path"
+  base_stream:
+    retriever:
+      $ref: "*ref(definitions.retriever)"
+      requester:
+        $ref: "*ref(definitions.requester)"
+  incremental_base_stream:
+    $ref: "*ref(definitions.base_stream)"
+    stream_cursor_field: "inserted_at"
+    retriever:
+      $ref: "*ref(definitions.retriever)"
+      requester:
+        $ref: "*ref(definitions.requester)"
+        request_options_provider:
+          request_parameters:
+            order_type: "desc"
+            order_by: "inserted_at"
+            date_shortcut: "custom"
+  surveys:
+    $ref: "*ref(definitions.base_stream)"
+    $options:
+      name: "surveys"
+      path: "surveys"
+      data_field: "surveys"
+  surveys_slicer:
+    class_name: source_zenloop.components.ZenloopSubstreamSlicer
+    $options:
+      config_parent_field: "survey_id"
+    parent_stream_configs:
+      - stream: "*ref(definitions.surveys)"
+        parent_key: public_hash_id
+        stream_slice_field: id
+  survey_groups:
+    $ref: "*ref(definitions.base_stream)"
+    $options:
+      name: "survey_groups"
+      path: "survey_groups"
+      data_field: "survey_groups"
+  survey_groups_slicer:
+    class_name: source_zenloop.components.ZenloopSubstreamSlicer
+    $options:
+      config_parent_field: "survey_group_id"
+    parent_stream_configs:
+      - stream: "*ref(definitions.survey_groups)"
+        parent_key: public_hash_id
+        stream_slice_field: id
+  date_slicer:
+    type: DatetimeStreamSlicer
+    cursor_field: "inserted_at"
+    datetime_format: "%Y-%m-%dT%H:%M:%S.%fZ"
+    start_datetime:
+      datetime: "{{ config['date_from'] }}"
+      datetime_format: "%Y-%m-%d"
+    end_datetime:
+      datetime: "{{ today_utc() }}"
+      datetime_format: "%Y-%m-%d"
+    step: "1m"
+    end_time_option:
+      field_name: "date_to"
+      inject_into: "request_parameter"
+    start_time_option:
+      field_name: "date_from"
+      inject_into: "request_parameter"
+  properties:
+    $ref: "*ref(definitions.base_stream)"
+    $options:
+      name: "properties"
+      data_field: "properties"
+    retriever:
+      $ref: "*ref(definitions.retriever)"
+      requester:
+        $ref: "*ref(definitions.requester)"
+        path: "{{ 'surveys/' + config['survey_id'] + '/properties' if config['survey_id'] else 'surveys/' + stream_slice.id + '/properties' }}"
+      stream_slicer:
+        $ref: "*ref(definitions.surveys_slicer)"
+  answers:
+    $ref: "*ref(definitions.incremental_base_stream)"
+    $options:
+      name: "answers"
+      data_field: "answers"
+    retriever:
+      $ref: "*ref(definitions.retriever)"
+      requester:
+        $ref: "*ref(definitions.incremental_base_stream.retriever.requester)"
+        path: "{{ 'surveys/' + stream_slice.id + '/answers' }}"
+      stream_slicer:
+        type: CartesianProductStreamSlicer
+        stream_slicers:
+          - "*ref(definitions.surveys_slicer)"
+          - "*ref(definitions.date_slicer)"
+  answers_survey_group:
+    $ref: "*ref(definitions.incremental_base_stream)"
+    $options:
+      name: "answers_survey_group"
+      data_field: "answers"
+    retriever:
+      $ref: "*ref(definitions.retriever)"
+      requester:
+        $ref: "*ref(definitions.incremental_base_stream.retriever.requester)"
+        path: "{{ 'survey_groups/' + stream_slice.id + '/answers' }}"
+      stream_slicer:
+        type: CartesianProductStreamSlicer
+        stream_slicers:
+          - "*ref(definitions.survey_groups_slicer)"
+          - "*ref(definitions.date_slicer)"
+
+
+streams:
+  - "*ref(definitions.surveys)"
+  - "*ref(definitions.survey_groups)"
+  - "*ref(definitions.properties)"
+  - "*ref(definitions.answers)"
+  - "*ref(definitions.answers_survey_group)"
+
+check:
+  type: CheckStream
+  stream_names: ["surveys"]
