--- airbyte-integrations/connectors/source-zenloop/unit_tests/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) 2021 Airbyte, Inc., all rights reserved.
-#
